TIQR-326: Fix for after removing an account, no account details can be opened
Ensure that the accounts details screen doesn't work with stale data: after the delete account operation is started (as a fire & forget), immediately close the detail screen. This ensures the details screen won't fire it's automatic close screen when the data is gone and get stuck there.
It gets stuck because the view model lifecycle is tied to the navigation graph and as long as the identities list screen remains open, the identities details view model will "remember" it's last identity details were null. Ergo, subsequently trying to open the details screen will immediately trigger closing the detail screen