TIQR-359: Refactor data module to be used directly in eduid (#82)
* TIQR-359: Refactor core for integrating into eduid
Multiple factors are now making it imperative for eduid to be dependent solely on the data module i.s.o. core module:
* to migrating to AGP 8+ and supporting R8 full mode
* replacing kapt with KSP
* removing redundant databinding for eduid
* migrating localisation to localicious
This will help make the integration of the tiqr core into eduid easier and it will be future compatible, i.s.o. still relying on deprecated/redundant or unused dependencies.
* Move shared strings from core to data module.
Use correct namespace for strings due to migration to AGP 8+
* Fix string namespaces due to AGP 8+ migration
* Fix migrating notification channel name
* Move camera dependency
* Fix packate declaration

---------

Co-authored-by: Iulia STANA <iulia@egeniq.com>